Add company name vector store setup script and update app to display short name
